Extract platform specific code to child classes
