+++ lib/defaultHttpClient.ts
@@ -1,4 +1,4 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
-export { FetchHttpClient as DefaultHttpClient } from "./fetchHttpClient";
+export { NodeFetchHttpClient as DefaultHttpClient } from "./nodeFetchHttpClient";
